# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TransEV EV CMS API
  version: 0.11.0
  description: |
    Canonical machine-readable contract for the endpoints currently implemented
    by `ev-cms-backend-new`.

    Authentication uses a signed-then-encrypted bearer access token. Platform
    sessions and CPO sessions are separate. CPO integration operations also
    require `X-CPO-App-ID`; that header is checked only after the bearer token
    establishes the CPO and never selects a tenant.

    Public customer signup requires `X-CPO-App-ID` to resolve an active CPO, but
    the header remains public routing metadata rather than authentication.
    Administrative authentication, platform control-plane, health, and
    documentation routes do not require it. Request bodies reject unknown
    properties and are limited to 32 KiB by the server.

    Every request that reaches Gin returns a server-generated UUID in the
    `X-Request-ID` response header for correlation with the safe structured
    completion log. The logger records matched route templates and stable
    handled error codes, never bodies, raw paths, queries, credentials, tokens,
    OTPs, email addresses, user agents, app IDs, API messages, or panic values.
    `LOG_LEVEL=DEBUG` adds request-start and handled-error component/type
    diagnostics under the same request ID without relaxing those exclusions.
  contact:
    name: TransEV Engineering
servers:
  - url: /
    description: Same origin as the API documentation page
...
paths:
...
  /api/v1/cpo/integrations:
    get:
      tags: [CPO Integrations]
      operationId: listCPOIntegrations
      summary: List configured integration metadata
      description: |
        Returns metadata only for the authenticated CPO. Secret plaintext is
        never returned. Requires the currently supported CPO ADMIN session, a changed initial
        password, and the current app-ID header.
      security:
        - BearerAuth: []
          CPOAppID: []
      responses:
        "200":
          description: Configured integration metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationListResponse"
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            `forbidden`, `password_change_required`, or `cpo_app_id_mismatch`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cpo/integrations/{provider}:
    parameters:
      - $ref: "#/components/parameters/IntegrationProvider"
    get:
      tags: [CPO Integrations]
      operationId: getCPOIntegration
      summary: Get one integration's metadata
      security:
        - BearerAuth: []
          CPOAppID: []
      responses:
        "200":
          description: Integration metadata without secret fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
        "400":
          description: "`missing_cpo_app_id` or `unsupported_integration_provider`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "404":
          $ref: "#/components/responses/IntegrationNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [CPO Integrations]
      operationId: putCPOIntegration
      summary: Configure or atomically rotate Razorpay credentials
      description: |
        Stores one encrypted credential object for the authenticated CPO.
        Associated authenticated-encryption data binds it to CPO and provider.
        The response and audit log contain metadata only.
      security:
        - BearerAuth: []
          CPOAppID: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RazorpayCredentialsRequest"
      responses:
        "200":
          description: Stored integration metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
        "400":
          description: |
            `invalid_request`, `missing_cpo_app_id`,
            `unsupported_integration_provider`, or
            `invalid_integration_credentials`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [CPO Integrations]
      operationId: deleteCPOIntegration
      summary: Remove the CPO's provider credentials
      description: Deletes the encrypted credential row and records a safe audit event.
      security:
        - BearerAuth: []
          CPOAppID: []
      responses:
        "204":
          description: Credentials removed. No response body.
        "400":
          description: "`missing_cpo_app_id` or `unsupported_integration_provider`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "404":
          $ref: "#/components/responses/IntegrationNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWE
      description: Short-lived signed-then-encrypted administrative access token.
    CPOAppID:
      type: apiKey
      in: header
      name: X-CPO-App-ID
      description: |
        Current app ID for the CPO already established by BearerAuth. This is
        routing metadata, not an authentication secret or tenant selector.
  parameters:
...
    IntegrationProvider:
      name: provider
      in: path
      required: true
      description: Provider name. Input is normalized to uppercase.
      schema:
        type: string
        enum: [RAZORPAY]
...
  responses:
...
    Unauthorized:
      description: "`unauthorized` — bearer token absent, invalid, expired, revoked, or no longer authorized."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: "`internal_error` — operation failed without exposing internals."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    MissingCPOAppID:
      description: "`missing_cpo_app_id` — X-CPO-App-ID is absent or blank."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    CPOIntegrationForbidden:
      description: |
        `forbidden`, `password_change_required`, or `cpo_app_id_mismatch`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    IntegrationNotFound:
      description: "`integration_not_found` — provider credentials are not configured."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    RazorpayCredentialsRequest:
      type: object
      additionalProperties: false
      required: [key_id, key_secret]
      properties:
        key_id:
          type: string
          minLength: 8
          maxLength: 100
          writeOnly: true
        key_secret:
          type: string
          minLength: 16
          maxLength: 255
          writeOnly: true
        webhook_secret:
          type: string
          maxLength: 255
          writeOnly: true
    Integration:
      type: object
      additionalProperties: false
      required:
        - provider
        - display_hint
        - is_active
        - configured_at
        - updated_at
      properties:
        provider:
          type: string
          enum: [RAZORPAY]
        display_hint:
          type: string
          description: Masked key-ID suffix, never a secret.
          example: "****5678"
        is_active:
          type: boolean
        configured_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IntegrationListResponse:
      type: object
      additionalProperties: false
      required: [integrations]
      properties:
        integrations:
          type: array
          items:
            $ref: "#/components/schemas/Integration"
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
            message:
              type: string
              description: Safe human-readable explanation.